feat: actions parameters
Add a top level 'actions' category, that lets you see the results of
actions that have been run, and a sub-application 'actions' that let you
see what actions you *could* run.

I'm not 100% happy that we are using actions to talk about the "shape of
the thing that could be run" and "the result of the thing that was run",
so we might want to use other names for these things.